Added an import for service request, so I could make an object of it to submit.
Changed naming of variable submitServiceRequest; previously ended with 'Service' which was unintended.

Finally, I implemented the handleServiceRequestSubmission method which creates an object of the ServiceRequest class in order to submit a service request.  This object is handled by the SubmitServiceRequestService class, which determines if it essential or non-essential, or if it's even a proper object.